+++ packages/side-panel/src/components/header/Component.tsx
@@ -57,6 +57,11 @@ export type HeaderProps = {
     sticky?: boolean;
 
     /**
+     * Фоновое изображение
+     */
+    imageUrl?: string;
+
+    /**
      * Идентификатор для систем автоматизированного тестирования
      */
     dataTestId?: string;
@@ -73,6 +78,7 @@ export const Header: FC<HeaderProps> = ({
     title,
     closer,
     sticky,
+    imageUrl,
     dataTestId,
 }) => {
     const { headerHighlighted, setHasHeader } = useContext(ModalContext);
@@ -86,11 +92,15 @@ export const Header: FC<HeaderProps> = ({
     return (
         <div
             className={cn(styles.header, className, {
+                [styles.backgroundImage]: imageUrl,
                 [styles.highlighted]: hasContent && sticky && headerHighlighted,
                 [styles.sticky]: sticky,
                 [styles.hasContent]: hasContent,
             })}
             data-test-id={dataTestId}
+            style={{
+                ...(imageUrl && { backgroundImage: `url(${imageUrl})` }),
+            }}
         >
             {leftAddons && <div className={cn(styles.addon, addonClassName)}>{leftAddons}</div>}